Move speaking module to user_interface subpackage
