+++ LTUAssistantPlus/user_interface/speaking.py
